+++ src/main/java/com/excelr/service/SubcategoryService.java
@@ -1,45 +1,71 @@
 package com.excelr.service;
 
-import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-
-import com.excelr.model.Product;
 import com.excelr.model.Subcategory;
 import com.excelr.repository.SubcategoryRepository;
 
 @Service
 public class SubcategoryService {
 	@Autowired
-	SubcategoryRepository subCategoriesRepo;
-	public List<Product> getProductsBySubCategoryId(Long subcategoryId){
-		Subcategory subCategories=subCategoriesRepo.findById(subcategoryId).orElseThrow(() -> new RuntimeException("Category not found"));
-		return subCategories.getProduct();
+	SubcategoryRepository  subCategoriesRepo;
+	public ResponseEntity<?> getProductsBySubCategoryId(Long subcategoryId){
+		Optional<Subcategory> subCategories=subCategoriesRepo.findById(subcategoryId);
+		if(subCategories.isPresent()) {
+			Subcategory subcategory=subCategories.get();
+			Map<String,Object> response=new HashMap<>();
+		return ResponseEntity.ok(response.put("subcategory", subcategory));
+		}else {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
 	}
-	public List<Subcategory> getSubCategories(){
-		return subCategoriesRepo.findAll();
+	public  ResponseEntity<?> getSubCategories(){
+		List<Subcategory> subcat= subCategoriesRepo.findAll();
+		Map<String, Object> res=new HashMap<>();
+		res.put("Subcategory", subcat);
+		return ResponseEntity.ok(res);
 	}
-	public Subcategory getSubCategoriesById(Long id) {
+	public ResponseEntity<?> getSubCategoriesById(Long id) {
         Optional<Subcategory> subcat= subCategoriesRepo.findById(id);
         if(subcat.isPresent()) {
-        	return subcat.get();
+        	Subcategory sub= subcat.get();
+        	Map<String, Object> res= new HashMap<>();
+        	res.put("SubCategory", sub);
+        	return ResponseEntity.ok(res);
         }else {
-        	return (Subcategory) Collections.emptyList();
+        	return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
-	public Subcategory createSubcategory(Subcategory subcategory) {
-        return subCategoriesRepo.save(subcategory);
+	public ResponseEntity<?> createSubcategory(Subcategory subcategory) {
+         subCategoriesRepo.save(subcategory);
+         return ResponseEntity.ok("saved successfully");
     }
-    public Subcategory updateSubcategory(Long id, Subcategory updatedSubcategory) {
-    	Subcategory subcategory = getSubCategoriesById(id);
-        subcategory.setName(updatedSubcategory.getName());
-        subcategory.setCategory(updatedSubcategory.getCategory());
-        return subCategoriesRepo.save(subcategory);
+    public ResponseEntity<?> updateSubcategory(Long id, Subcategory updatedSubcategory) {
+    	Optional<Subcategory> subcate = subCategoriesRepo.findById(id);
+    	if(subcate.isPresent()) {
+    		Subcategory subcategory= subcate.get();
+    		subcategory.setName(updatedSubcategory.getName());
+            subcategory.setCategory(updatedSubcategory.getCategory());
+            subCategoriesRepo.save(subcategory);
+            Map <String,String> res=new HashMap<> ();
+            return ResponseEntity.ok(res.put("message", "updated successfully"));
+    	}else {
+    		return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    	}
     }
-    public void deleteSubcategory(Long id) {
+    public ResponseEntity<?> deleteSubcategory(Long id) {
+    	Optional<Subcategory> sub=subCategoriesRepo.findById(id);
+    	if(sub.isPresent()) {
     	subCategoriesRepo.deleteById(id);
+    	return ResponseEntity.ok("deleted successfully");
+    	}else {
+    	return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    	}
     }
 }
